#37 references being lost during migration.
References are lost by the child folder on rename.  Presumably because it doesn't extend from BaseFolder and so misses the important call to BaseObject._notifyOfCopyTo.

Fix by not renaming the child folder, instead create new folder with temp id and rename that to the child folder id after the child folder has been deleted

http://plone.org/products/collective-lineage/issues/37